requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
atoma==0.0.15
attrs==19.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
boto3==1.9.125
botocore==1.12.125
cached-property==1.5.1
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
Click==7.0
cryptography==2.5
cycler==0.10.0
defusedxml==0.6.0
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-MySQL==1.4.0
gast==0.2.2
gcloud==0.18.3
google==2.0.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-automl==0.1.2
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.1
idna==2.7
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mathparse==0.1.2
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
mysql-connector-python==8.0.16
nltk==3.4
numpy==1.16.3
oauth2client==4.1.3
pandas==0.24.2
paramiko==2.4.2
pbr==5.1.2
Pint==0.9
protobuf==3.6.1
protobuf3-to-dict==0.1.5
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pylint==2.3.1
pymongo==3.7.2
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.20.1
rsa==4.0
s3transfer==0.2.0
sagemaker==1.18.9
scikit-learn==0.19.2
scipy==1.2.1
seaborn==0.9.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
soupsieve==1.8
SQLAlchemy==1.2.18
storage==0.0.4.3
tensorboard==1.12.2
tensorflow==1.13.0rc2
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
texttable==0.9.1
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.4.0
websocket-client==0.56.0
Werkzeug==0.14.1
wrapt==1.11.1
Flask-Cors==3.0.7 
